# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: UserResponse API Backend
  description: API documentation for managing UserResponses, including fetching by ID, email, and performing full-text searches.
  version: 1.0.0
servers:
  - url: /api/users
paths:
  /{id}:
    get:
      tags:
        - User
      summary: Get UserResponse by ID
      description: Retrieves a UserResponse by their unique ID.
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the UserResponse.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: UserResponse found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: UserResponse not found

  /email:
    get:
      tags:
        - User
      summary: Get UserResponse by email
      description: Retrieves a UserResponse by their email address.
      operationId: getUserByEmail
      parameters:
        - name: email
          in: query
          required: true
          description: Email address of the UserResponse.
          schema:
            type: string
            format: email
      responses:
        '200':
          description: UserResponse found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid email format
        '404':
          description: UserResponse not found

  /search:
    get:
      tags:
        - User
      summary: Search UserResponses
      description: Performs a full-text search on UserResponses using a keyword.
      operationId: searchUser
      parameters:
        - name: keyword
          in: query
          required: true
          description: Keyword to search for (matches firstName, lastName, and ssn).
          schema:
            type: string
      responses:
        '200':
          description: Search successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '404':
          description: No UserResponses found

components:
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        age:
          type: integer
          example: 30
        gender:
          type: string
          example: male
        email:
          type: string
          format: email
          example: john.doe@example.com
        phone:
          type: string
          example: "+1 555-1234"
        birthDate:
          type: string
          format: date
          example: 1993-01-15
        ssn:
          type: string
          example: "123-45-6789"
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  responses:
    UserResponseNotFound:
      description: UserResponse not found.
    InvalidEmailFormat:
      description: Invalid email format.
    NoUserResponsesFound:
      description: No UserResponses found for the given keyword.

security:
  - basicAuth: []